pyproject.toml
[project]
name = "avd-to-markdown"
version = "0.1.0"
description = "Takes avd to markdown within mermaid"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "pyaml>=25.1.0",
]
